pyproject.toml
[project]
name = "fiveaxissimcore"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cupy-cuda12x>=13.6.0",
    "networkx>=3.5",
    "numpy>=2.3.3",
    "open3d>=0.19.0",
    "pyqt6>=6.10.0",
    "pyvista>=0.46.3",
    "rtree>=1.4.1",
    "scipy>=1.16.2",
    "tqdm>=4.67.1",
    "trimesh>=4.8.3",
]
